Add AuthCubit and update SignUpScreen with BlocProvider
- Added AuthCubit to manage authentication state
- Updated SignUpScreen to use BlocProvider for AuthCubit
- Created CustomHaveAccount widget for reusability
