+++ lib/core/routes/app_router.dart
@@ -1,8 +1,11 @@
 import 'package:dalel/core/routes/routes.dart';
+import 'package:dalel/core/services/service_locator.dart';
+import 'package:dalel/features/auth/auth_cubit/cubit/auth_cubit.dart';
 import 'package:dalel/features/auth/presentation/screens/sign_in_screen.dart';
 import 'package:dalel/features/auth/presentation/screens/sign_up_screen.dart';
 import 'package:dalel/features/on_Boarding/presentation/screens/on_boarding_screen.dart';
 import 'package:dalel/features/splash/presentation/splash_screen.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
 
 final GoRouter appRouter = GoRouter(
@@ -17,11 +20,17 @@ final GoRouter appRouter = GoRouter(
     ),
     GoRoute(
       path: signUpScreen,
-      builder: (context, state) => const SignUpScreen(),
+      builder: (context, state) => BlocProvider(
+        create: (context) => getIt<AuthCubit>(),
+        child: const SignUpScreen(),
+      ),
     ),
     GoRoute(
       path: signInScreen,
-      builder: (context, state) => const SignInScreen(),
+      builder: (context, state) => BlocProvider(
+        create: (context) => getIt<AuthCubit>(),
+        child: const SignInScreen(),
+      ),
     ),
   ],
 );
