+++ lib/core/services/service_locator.dart
@@ -1,8 +1,10 @@
 import 'package:dalel/core/database/cache/cache_helper.dart';
+import 'package:dalel/features/auth/auth_cubit/cubit/auth_cubit.dart';
 import 'package:get_it/get_it.dart';
 
 final getIt = GetIt.instance;
 
 void setupServiceLocator() {
   getIt.registerSingleton<CacheHelper>(CacheHelper());
+  getIt.registerSingleton<AuthCubit>(AuthCubit());
 }
