+++ lib/features/auth/auth_cubit/cubit/auth_state.dart
@@ -1,0 +1,3 @@
+ class AuthState {}
+
+final class AuthInitial extends AuthState {}
