+++ lib/features/auth/presentation/widgets/have_account_widget.dart
@@ -1,0 +1,26 @@
+import 'package:dalel/core/utils/app_text_styles.dart';
+import 'package:flutter/material.dart';
+
+class CustomHaveAccount extends StatelessWidget {
+  const CustomHaveAccount(
+      {super.key, required this.text1, required this.text2});
+  final String text1, text2;
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text.rich(
+        TextSpan(
+          text: text1,
+          style: CustomTextStyles.poppins400style12,
+          children: [
+            TextSpan(
+              text: text2,
+              style: CustomTextStyles.poppins500style12
+                  .copyWith(decoration: TextDecoration.underline),
+            )
+          ],
+        ),
+      ),
+    );
+  }
+}
