+++ pubspec.yaml
@@ -28,9 +28,11 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
+  bloc: ^8.1.4
   cupertino_icons: ^1.0.6
   flutter:
     sdk: flutter
+  flutter_bloc: ^8.1.5
   flutter_native_splash: ^2.4.0
   flutter_screenutil: ^5.9.0
   gap: ^3.0.1
